requirements.txt
streamlit
pandas>=2.2.3
requests
google-genai
spacy
watchdog